requirements.txt
# requirements.txt
aiohttp
asyncio
beautifulsoup4
langchain
openai
requests
tqdm
spacy
python-dotenv
nltk
numpy